support for piped HTML data
This allows for piped data to be HTML and passed through to the
matrix message rather than adding the `<code>` formatting as was
default before.

I needed this because I wanted to send hyperlinks and thus allowing
the sending program to output HTML was the easiest thing to do here.